can create non-existing file
do not respect the line numbers for the moment

+++ src/fixlines.cr
@@ -10,7 +10,11 @@ ARGF.each_line { |change|
 }
 
 seen.each { |path,changes|
-	lines = File.read_lines path
-	changes.each { |no,content| lines[no-1] = content }
-	lines = File.write(path, (lines.join "\n")+"\n")
+	if File.exists? path
+		lines = File.read_lines path
+		changes.each { |no,content| lines[no-1] = content }
+		File.write(path, (lines.join "\n")+"\n")
+	else
+		File.write(path, (changes.map { |_,content| content }.join "\n")+"\n")
+	end
 }
